backend/requirements.txt
mysql-connector-python
flask
flask-cors
gunicorn